==> notebooks/02_fetch_index_prices.ipynb <==
import yfinance as yf
import pandas as pd

djia = yf.download("^DJI", start="2008-08-01", end="2017-01-01")
djia.reset_index(inplace=True)
djia.head()

djia = djia.rename(columns={"Date": "trading_date", "Close": "close"})

djia = djia[["trading_date", "close", "Volume"]]

djia.to_parquet("data/raw/djia_prices.parquet")
